Add expectation code
Simple expectation functions to be used in test suites. These functions
provide a simple interface to check for equality of both simple values
and of stream. The latter uses reflection to determine the type of the
expected value in order to know how to properly decode the stream.

In a future version of Go (possible 1.17) these function may be
rewritten using generics to remove the need for reflection and interface
types.